# api/openapi.yaml
openapi: "3.1.0"
info:
  title: "Kinetic v3 API"
  version: "0.1.0"
  description: |
    REST API for the Kinetic v3 Rocket League esports league management platform.

    ## Authentication
    Most read endpoints are publicly accessible. Write and sensitive operations
    require authentication via either a session cookie (set by the OAuth login flow)
    or an HTTP Bearer token (API token issued via `/v1/api-tokens`).

    ## Permission model
    Protected routes enforce a resource/action pair checked against the caller's
    roles. Available resources are defined in `authz.resources` and include
    `league`, `franchise`, `club`, `team`, `player`, `roster_membership`,
    `role_assignment`, `queue`, `queue_entry`, `queue_ban`, `scrim`,
    `scrim_promotion`, `result_submission`, `result_override`, `replay_evidence`,
    `platform_account_link`, `player_rating`, `rating_adjustment`,
    `matchmaking_decision`, `exception_ticket`, `season`, `schedule_group`,
    `fixture`, `match`, `eligibility`, `player_stat`, `skill_group`,
    `skill_group_transition`, `organization_config`, `api_token`, and
    `player_notification`.
servers:
  - url: "http://localhost:8080"
    description: "Local development server"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: kinetic_session
      description: "Signed session token set by the OAuth login flow."
    bearerToken:
      type: http
      scheme: bearer
      description: "Long-lived API token issued via POST /v1/api-tokens."
  schemas:
...
    ResultSubmission:
      type: object
      properties:
        id:
          type: integer
          format: int64
        contextType:
          type: string
          description: "Either 'scrim' or 'match'."
        contextId:
          type: integer
          format: int64
        submittedByTeamId:
          type: integer
          format: int64
        homeTeamId:
          type: integer
          format: int64
        awayTeamId:
          type: integer
          format: int64
        winningTeamId:
          type: integer
          format: int64
        losingTeamId:
          type: integer
          format: int64
        state:
          type: string
          description: "Workflow state (e.g. pending, ratified, rejected, overridden)."
        payloadJson:
          description: "Arbitrary JSON payload attached to the submission."
        homeRatifiedAt:
          type: string
          format: date-time
          nullable: true
        awayRatifiedAt:
          type: string
          format: date-time
          nullable: true
        rejectedByTeamId:
          type: integer
          format: int64
          nullable: true
        rejectionReason:
          type: string
          nullable: true
        rejectedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - contextType
        - contextId
        - submittedByTeamId
        - homeTeamId
        - awayTeamId
        - winningTeamId
        - losingTeamId
        - state
        - createdAt
    ResultOverride:
      type: object
      properties:
        id:
          type: integer
          format: int64
        submissionId:
          type: integer
          format: int64
        actor:
          type: string
        reason:
          type: string
        previousWinningTeamId:
          type: integer
          format: int64
        previousLosingTeamId:
          type: integer
          format: int64
        newWinningTeamId:
          type: integer
          format: int64
        newLosingTeamId:
          type: integer
          format: int64
        previousState:
          type: string
        newState:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - submissionId
        - actor
        - reason
        - previousWinningTeamId
        - previousLosingTeamId
        - newWinningTeamId
        - newLosingTeamId
        - previousState
        - newState
        - createdAt

    # ── Replay ─────────────────────────────────────────────────────────────────
    ReplayEvidence:
      type: object
      properties:
        id:
          type: integer
          format: int64
        contextType:
          type: string
        contextId:
          type: integer
          format: int64
        submittedByTeamId:
          type: integer
          format: int64
        replaySha256:
          type: string
        contentSizeBytes:
          type: integer
          format: int64
        storageRef:
          type: string
        state:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - contextType
        - contextId
        - submittedByTeamId
        - replaySha256
        - contentSizeBytes
        - storageRef
        - state
        - createdAt
    ReplayParseRun:
      type: object
      properties:
        id:
          type: integer
          format: int64
        replayEvidenceId:
          type: integer
          format: int64
        parserName:
          type: string
        parserVersion:
          type: string
        parserConfigDigest:
          type: string
        status:
          type: string
        outputJson:
          description: "Raw parsed output from the replay parser."
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - replayEvidenceId
        - parserName
        - parserVersion
        - parserConfigDigest
        - status
        - createdAt
    ReplayIngestionResult:
      type: object
      properties:
        evidence:
          $ref: "#/components/schemas/ReplayEvidence"
        parseRun:
          $ref: "#/components/schemas/ReplayParseRun"
        duplicate:
          type: boolean
        linkedSubmissionId:
          type: integer
          format: int64
          nullable: true
      required:
        - evidence
        - parseRun
        - duplicate
    ResultSubmissionReplayLink:
      type: object
      properties:
        id:
          type: integer
          format: int64
        resultSubmissionId:
          type: integer
          format: int64
        replayEvidenceId:
          type: integer
          format: int64
        linkedByTeamId:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - resultSubmissionId
        - replayEvidenceId
        - linkedByTeamId
        - createdAt

    # ── Stats ──────────────────────────────────────────────────────────────────
    PlayerStatLine:
      type: object
      properties:
        id:
          type: integer
          format: int64
        roundId:
          type: integer
          format: int64
        playerId:
          type: integer
          format: int64
          nullable: true
        replayIdentity:
          type: string
        goals:
          type: integer
          format: int32
        assists:
          type: integer
          format: int32
        saves:
          type: integer
          format: int32
        shots:
          type: integer
          format: int32
        score:
          type: integer
          format: int32
        opi:
          type: number
          format: double
          description: "Offensive performance index."
        dpi:
          type: number
          format: double
          description: "Defensive performance index."
        gpi:
          type: number
          format: double
          description: "General performance index (opi + dpi)."
      required:
        - id
        - roundId
        - replayIdentity
        - goals
        - assists
        - saves
        - shots
        - score
        - opi
        - dpi
        - gpi
    PlayerCareerStats:
      type: object
      properties:
        playerId:
          type: integer
          format: int64
        totalGames:
          type: integer
          format: int32
        totalGoals:
          type: integer
          format: int32
        totalAssists:
          type: integer
          format: int32
        totalSaves:
          type: integer
          format: int32
        totalShots:
          type: integer
          format: int32
        avgOpi:
          type: number
          format: double
        avgDpi:
          type: number
          format: double
        avgGpi:
          type: number
          format: double
      required:
        - playerId
        - totalGames
        - totalGoals
        - totalAssists
        - totalSaves
        - totalShots
        - avgOpi
        - avgDpi
        - avgGpi

    # ── Exceptions ─────────────────────────────────────────────────────────────
...
paths:
...
  /v1/players/{id}/career-stats:
    get:
      operationId: getPlayerCareerStats
      summary: "Get career-aggregate stats for a player"
      tags:
        - replay
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "Career stats."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerCareerStats"
        "400":
          description: "Invalid player ID."
        "503":
          description: "Replay stats store unavailable."
...
  /v1/result-submissions:
    get:
      operationId: listResultSubmissions
      summary: "List all result submissions"
      tags:
        - results
      responses:
        "200":
          description: "All result submissions."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResultSubmission"
        "503":
          description: "Hierarchy store unavailable."
    post:
      operationId: createResultSubmission
      summary: "Submit a match result"
      tags:
        - results
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                contextType:
                  type: string
                  description: "Either 'scrim' or 'match'."
                contextId:
                  type: integer
                  format: int64
                submittedByTeamId:
                  type: integer
                  format: int64
                winningTeamId:
                  type: integer
                  format: int64
                losingTeamId:
                  type: integer
                  format: int64
                payloadJson:
                  description: "Arbitrary JSON payload."
              required:
                - contextType
                - contextId
                - submittedByTeamId
                - winningTeamId
                - losingTeamId
      responses:
        "201":
          description: "Created result submission."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultSubmission"
        "400":
          description: "Invalid input."
        "409":
          description: "Conflict."
        "503":
          description: "Hierarchy store unavailable."
...
  /v1/result-submissions/{id}/stats:
    get:
      operationId: listResultSubmissionStats
      summary: "List player stat lines for a result submission"
      tags:
        - replay
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "Player stat lines."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlayerStatLine"
        "400":
          description: "Invalid submission ID."
        "503":
          description: "Replay stats store unavailable."
  /v1/result-submission-ratifications:
    post:
      operationId: ratifyResultSubmission
      summary: "Ratify a result submission on behalf of a team"
      tags:
        - results
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                submissionId:
                  type: integer
                  format: int64
                teamId:
                  type: integer
                  format: int64
              required:
                - submissionId
                - teamId
      responses:
        "200":
          description: "Updated result submission."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultSubmission"
        "400":
          description: "Invalid input."
        "409":
          description: "Already ratified or in wrong state."
        "503":
          description: "Hierarchy store unavailable."
  /v1/result-submission-rejections:
    post:
      operationId: rejectResultSubmission
      summary: "Reject a result submission on behalf of a team"
      tags:
        - results
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                submissionId:
                  type: integer
                  format: int64
                teamId:
                  type: integer
                  format: int64
                reason:
                  type: string
              required:
                - submissionId
                - teamId
                - reason
      responses:
        "200":
          description: "Updated result submission."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultSubmission"
        "400":
          description: "Invalid input."
        "409":
          description: "Already rejected or in wrong state."
        "503":
          description: "Hierarchy store unavailable."
  /v1/result-overrides:
    get:
      operationId: listResultOverrides
      summary: "List all result overrides"
      tags:
        - results
      responses:
        "200":
          description: "All result overrides."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResultOverride"
        "503":
          description: "Hierarchy store unavailable."
    post:
      operationId: overrideResultSubmission
      summary: "Override the outcome of a result submission (admin)"
      tags:
        - results
      security:
        - sessionCookie: []
        - bearerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                submissionId:
                  type: integer
                  format: int64
                actor:
                  type: string
                reason:
                  type: string
                winningTeamId:
                  type: integer
                  format: int64
                losingTeamId:
                  type: integer
                  format: int64
              required:
                - submissionId
                - actor
                - reason
                - winningTeamId
                - losingTeamId
      responses:
        "200":
          description: "Updated result submission."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultSubmission"
        "400":
          description: "Invalid input."
        "401":
          description: "Not authenticated."
        "403":
          description: "Insufficient permissions (result_override:create required)."
        "503":
          description: "Hierarchy store unavailable."

  # ── Replay ─────────────────────────────────────────────────────────────────
  /v1/replay-evidence:
    get:
      operationId: listReplayEvidence
      summary: "List all replay evidence records"
      tags:
        - replay
      responses:
        "200":
          description: "All replay evidence."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReplayEvidence"
        "503":
          description: "Hierarchy store unavailable."
    post:
      operationId: ingestReplayEvidence
      summary: "Ingest a replay file and trigger parse"
      tags:
        - replay
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                contextType:
                  type: string
                contextId:
                  type: integer
                  format: int64
                submittedByTeamId:
                  type: integer
                  format: int64
                replayBody:
                  type: string
                  description: "Base64-encoded or raw replay file content."
                parserName:
                  type: string
                parserVersion:
                  type: string
                parserConfigDigest:
                  type: string
                parseOutputJson:
                  description: "Pre-parsed output JSON if available."
                resultSubmissionId:
                  type: integer
                  format: int64
                  nullable: true
              required:
                - contextType
                - contextId
                - submittedByTeamId
                - replayBody
                - parserName
                - parserVersion
                - parserConfigDigest
      responses:
        "201":
          description: "Replay ingested (new)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplayIngestionResult"
        "200":
          description: "Replay already known (duplicate SHA256)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplayIngestionResult"
        "400":
          description: "Invalid input."
        "503":
          description: "Hierarchy store unavailable."
  /v1/replay-parse-runs:
    get:
      operationId: listReplayParseRuns
      summary: "List all replay parse runs"
      tags:
        - replay
      responses:
        "200":
          description: "All replay parse runs."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReplayParseRun"
        "503":
          description: "Hierarchy store unavailable."
  /v1/result-submission-replay-links:
    get:
      operationId: listResultSubmissionReplayLinks
      summary: "List all replay links attached to result submissions"
      tags:
        - replay
      responses:
        "200":
          description: "All replay links."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResultSubmissionReplayLink"
        "503":
          description: "Hierarchy store unavailable."

  # ── Exceptions ─────────────────────────────────────────────────────────────
...